+++ apps/api/src/app/symbol/symbol.controller.ts
@@ -1,4 +1,4 @@
-import { RequestWithUser } from '@ghostfolio/helper/types';
+import { RequestWithUser } from '@ghostfolio/common/types';
 import {
   Controller,
   Get,