+++ apps/api/src/models/interfaces/rule.interface.ts
@@ -1,4 +1,4 @@
-import { PortfolioPosition } from '@ghostfolio/helper/interfaces';
+import { PortfolioPosition } from '@ghostfolio/common/interfaces';
 
 import { EvaluationResult } from './evaluation-result.interface';