+++ apps/api/src/services/data-provider/rakuten-rapid-api/rakuten-rapid-api.service.ts
@@ -1,5 +1,5 @@
-import { getToday, getYesterday } from '@ghostfolio/helper/helper';
-import { Granularity } from '@ghostfolio/helper/types';
+import { getToday, getYesterday } from '@ghostfolio/common/helper';
+import { Granularity } from '@ghostfolio/common/types';
 import { Injectable } from '@nestjs/common';
 import { DataSource } from '@prisma/client';
 import * as bent from 'bent';