+++ apps/client/src/app/components/value/value.component.ts
@@ -5,7 +5,7 @@ import {
   OnChanges,
   OnInit
 } from '@angular/core';
-import { DEFAULT_DATE_FORMAT } from '@ghostfolio/helper/config';
+import { DEFAULT_DATE_FORMAT } from '@ghostfolio/common/config';
 import { format, isDate } from 'date-fns';
 import { isNumber } from 'lodash';